Move artist property to Basic Object
upnp devices often return the artist property for the
`object.container.album.musicAlbum` class. By
restricting the property to the item level
(`object.item.audioItem.musicTrack`), it makes it awkward to
access this property from the music album folder level.

This patch moves the property up to the Basic Object level so it may be
used earlier.