Add support for multiple resource objects
Previously the parser would set attributes based on the last resource
object in the resource array. This was problematic with certain older
UPnP servers which declared thumbnails as resource object.

This patch means that uri, protocol info and other necessary attributes
are set by the LAST resource object providing the protocol does not
contain the string "image/".